+++ test/readdir_inode.c
@@ -1,7 +1,8 @@
 /*
- * Prints each directory entry and its inode as returned by 'readdir'.
+ * Prints each directory entry, its inode and d_type as returned by 'readdir'.
  * Skips '.' and '..' because readdir is not required to return them and
- * some of our examples don't.
+ * some of our examples don't. However if they are returned, their d_type
+ * should be valid.
  */
 
 #include <stdio.h>
@@ -30,7 +31,18 @@ int main(int argc, char* argv[])
     dent = readdir(dirp);
     while (dent != NULL) {
         if (strcmp(dent->d_name, ".") != 0 && strcmp(dent->d_name, "..") != 0) {
-            printf("%llu %s\n", (unsigned long long)dent->d_ino, dent->d_name);
+            printf("%llu %d %s\n", (unsigned long long)dent->d_ino,
+			(int)dent->d_type, dent->d_name);
+            if ((long long)dent->d_ino < 0)
+               fprintf(stderr,"%s : bad d_ino %llu\n",
+                        dent->d_name, (unsigned long long)dent->d_ino);
+            if ((dent->d_type < 1) || (dent->d_type > 15))
+               fprintf(stderr,"%s : bad d_type %d\n",
+                        dent->d_name, (int)dent->d_type);
+        } else {
+            if (dent->d_type != DT_DIR)
+               fprintf(stderr,"%s : bad d_type %d\n",
+                        dent->d_name, (int)dent->d_type);
         }
         dent = readdir(dirp);
     }